+++ src/components/Select/theme/index.js
@@ -20,6 +20,7 @@ export default ({ colors, forms, sizes, spaces }) => ({
       arrowWidth: '10px',
       minHeight: forms.sizes.minHeight,
       separatorWidth: sizes.widths.w1,
+      width: sizes.widths.full,
     },
     spaces: {
       paddingHeight: spaces.s5,